Adding aliases for grep, screen and tmux.
For grep it provides easier recursive command "rgrep".
For screen it provides easier listing and reattaching command "sdr".
For tmux, "jtmux" is a wrapper around the original command providing easier new
and attach functionality. It also provides "revive" which attaches and refreshes
the SSH_AUTH_SOCK environmental variable for all windows in the given session.
That is useful for tmux sessions that span multiple ssh connections.