[alias] Adding alias less-log.
The underlying function requires two argumnets: i) the directory
containing the log files, and ii) a file name pattern to search for
within that directory. For the files matching the pattern it sorts
them by modification date and opens the first in that sorted
list (newest) with `poinitng` to the end of the file.