[git] Adding workflow alias `j-restore-all-but` and introducing the `.gitconfig-aliases.sh` file.
This workflow restores all the files in the current working
directory. It gets a list of files as arguments which excludes from
the files it restores. This workflow helps in cases when you
accidentally change a lot of files but you cannot restore the whole
working directory because some of the files contain changes you wish
to preserve. In that case you need to be more surgical and be able to
restore only the files you want. For example imagine you have been
working on a feature that touched 4 files but at some point you
accidentally execute your project's linter with a faulty configuration
which in turn touches 150 files of your project; with this workflow
you can restore all the files except from the 4 files you were working
on.
This change also introduces the `.gitconfig-aliases.sh` file which
will contain all the shell functions that implement the more complex
git workflows/aliases improving readability and maintainability.